Adds `next_update_at` column to Player model.
Is used to indicate when the next update will be needed. This removes
some logic from StalerPlayersUpdate, and pushes it into the Player
model.

It is also more robust than matching on Player.updated_at, as that one
can get updated for reasons other than new data - e.g. if the
`update_scheduled_at` timestamp is cleared.