Fix tagging of container version in Makefile.
The container was tagged with the commit at the time of calling `make` -
which did not contain potential commits, with changes to version and
changelog, created by the `tag` target.

While not affecting the code within the container, it did mean that, in
practice, the commit tags of Docker containers were always one off,
causing `util/push_containers` to fail to auto-tag as named versions.