Add caching of player ranks.
Ranks for Hive skill, experience etc are periodically calculated and
cached in Redis, from where the view retrieves it.

This prevents a blocking query on each page load which, despite heavy
optimization, takes multiple seconds.

Due to the nature of rank calculation, it takes no more effort to
calculate all ranks than a single one, making it a good fit for caching.