Strip query parameters whose value is nil, fixing an exception.
If a parameter is given as `/foo?bar` - ie without an equals sign - it
ends up in the `params` hash with value `nil`.

This caused an exception in player search, due to it only checking the
existence of the key, and assuming its value to be, at the very least,
an empty string.

As I do not use query parameters as toggles, this commit strips them
completely.

Also refactors controller action to not use exception for control flow.

Resolves #33.